Ajustes para iniciar teste
