+++ main.py
@@ -2,6 +2,7 @@ from services.validate import ValidateToken
 from routers import clients, favorites
 from dependencies import users_db, authenticate_user, create_access_token, ACCESS_TOKEN_EXPIRE_MINUTES, get_current_user
 
+import uvicorn
 from fastapi.security.oauth2 import OAuth2PasswordRequestForm
 from datetime import timedelta
 from fastapi import Depends, HTTPException, status, FastAPI
@@ -56,3 +57,6 @@ def rotas():
     app.include_router(clients.router)
     app.include_router(favorites.router)
 
+if __name__ == '__main__':
+    rotas()
+    uvicorn.run(app, port=8000, host='0.0.0.0')
